requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.8.0
dash-bootstrap-components==1.3.0
dash-bootstrap-templates==1.0.7
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
fake-useragent==1.1.1
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Levenshtein==0.20.9
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
rapidfuzz==2.13.7
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.1.0
tqdm==4.64.1
transliterate==1.10.2
urllib3==1.25.11
Werkzeug==2.2.2
wordcloud==1.8.2.2
yarl==1.8.2
